requirements.txt
Flask==3.0.3
psycopg2==2.9.7
sentence-transformers==3.3.1
groq==0.13.0
python-dotenv==1.0.0
FlagEmbedding==1.3.3
